requirements.txt
flask>=3.0,<4.0
flask-cors>=4.0,<5.0
werkzeug>=3.0,<4.0
requests>=2.31,<3.0
chromadb>=0.5.5,<1.0
pandas>=2.0,<3.0
openpyxl>=3.1,<4.0
pyyaml>=6.0,<7.0
python-docx>=1.1,<2.0
pymupdf>=1.24,<2.0
pillow>=10.0,<12.0
pytesseract>=0.3.10,<1.0
websockets>=12.0,<16.0
beautifulsoup4>=4.12,<5.0
fastapi>=0.110,<1.0
uvicorn>=0.30,<1.0
python-dotenv>=1.0,<2.0
